puszczenie prettiera i przeniesienie typu, stary typesript 4.0 nie osbługuje takiego importu typów
